see #48
